pyproject.toml
[project]
name = "railmind"
version = "0.1.0"
description = "RailMind — AI Intelligence Layer for Mumbai Suburban Railways"
requires-python = ">=3.12"
dependencies = [
    # Core API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",

    # Neo4j Knowledge Graph
    "neo4j>=5.25.0",

    # LLM & Embeddings
    "openai>=1.50.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.28",

    # Milvus (Dhruv - RAG vector store)
    "pymilvus>=2.4.7",

    # NLP Pipeline (CrowdSignal)
    "spacy>=3.7.6",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.3",

    # LangChain Core
    "langchain-core>=0.3.0",
    "tiktoken>=0.7.0",
    "httpx>=0.27.0",

    # Scheduling & Notifications
    "apscheduler>=3.10.4",

    # Weather (uses requests)
]
